move wcs_coordinator/waste_collection_api in own file + readd fetch service for UI config
